e2e-tests folder organisation
