+++ e2e-tests/actions/watchActiveAccount.ts
@@ -19,7 +19,7 @@
 
 import {Page} from '@playwright/test';
 
-import {App} from './createApp';
+import {App} from '../types';
 
 // eslint-disable-next-line valid-jsdoc
 /**
