+++ e2e-tests/backend/PublicApiClient.ts
@@ -21,11 +21,7 @@ import {ok, type RequestOpts} from '@oazapfts/runtime';
 
 import * as publicApiClient from './generated/publicApi';
 
-import {User} from '../actions/createUser';
-
-export type RegisteredUser = User & {id: string; token: string};
-
-export type TeamOwner = RegisteredUser & {teamId: string};
+import {User, RegisteredUser, TeamOwner} from '../types';
 
 export type PublicApiClientConfig = {
   baseUrl: string;
